comment 타입 정리
